backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.1",
    "pydantic-settings>=2.14.0",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.46.0",
    "pydantic[email]>=2.13.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "stripe[async]>=15.1.0",
]
